Moved remaining character related handlers to the character service.
