+++ src/swganh/character/base_character_service.h
@@ -27,7 +27,7 @@
 
 #include "swganh/base/base_service.h"
 #include "swganh/character/character_data.h"
-#include "connection/messages/client_create_character.h"
+#include "swganh/character/messages/client_create_character.h"
 
 namespace anh {
 namespace app {
@@ -46,7 +46,7 @@ public:
     virtual CharacterLoginData GetLoginCharacter(uint64_t character_id, uint64_t account_id) = 0;
     virtual bool DeleteCharacter(uint64_t character_id, uint64_t account_id) = 0;
     virtual std::wstring GetRandomNameRequest(const std::string& base_model) = 0;
-    virtual std::tuple<uint64_t, std::string> CreateCharacter(const connection::messages::ClientCreateCharacter& character_info, uint32_t account_id) = 0;
+    virtual std::tuple<uint64_t, std::string> CreateCharacter(const swganh::character::messages::ClientCreateCharacter& character_info, uint32_t account_id) = 0;
     virtual uint16_t GetMaxCharacters(uint64_t player_id) = 0;
 
 private:
