+++ src/swganh/character/messages/client_create_character_success.h
@@ -18,31 +18,32 @@
  Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
 */
 
-#ifndef CONNECTION_MESSAGES_CLIENT_RANDOM_NAME_REQUEST_H_
-#define CONNECTION_MESSAGES_CLIENT_RANDOM_NAME_REQUEST_H_
+#ifndef SWGANH_CHARACTER_MESSAGES_CLIENT_CHARACTER_CREATE_SUCCESS_H_
+#define SWGANH_CHARACTER_MESSAGES_CLIENT_CHARACTER_CREATE_SUCCESS_H_
 
 #include <cstdint>
 #include "anh/byte_buffer.h"
 #include "swganh/base/swg_message.h"
 
-namespace connection {
+namespace swganh {
+namespace character {
 namespace messages {
     
-struct ClientRandomNameRequest : public swganh::base::SwgMessage<ClientRandomNameRequest> {
+struct ClientCreateCharacterSuccess : public swganh::base::SwgMessage<ClientCreateCharacterSuccess> {
     static const uint16_t opcount = 2;
-    static const uint32_t opcode = 0xD6D1B6D1;    
+    static const uint32_t opcode = 0x1DB575CC;    
     
-    std::string player_race_iff;
+    uint64_t character_id;
 
     void onSerialize(anh::ByteBuffer& buffer) const {
-        buffer.write(player_race_iff);	
+        buffer.write(character_id);	
     }
 
     void onDeserialize(anh::ByteBuffer buffer) {
-    	player_race_iff = buffer.read<std::string>();
+    	character_id = buffer.read<uint64_t>();
     }
 };
 
-}}  // namespace connection::messages
+}}}  // namespace swganh::character::messages
 
-#endif  // CONNECTION_MESSAGES_CLIENT_RANDOM_NAME_REQUEST_H_
+#endif  // SWGANH_CHARACTER_MESSAGES_CLIENT_CHARACTER_CREATE_SUCCESS_H_
